修改部分事件名及参数数据类型；修改部分信息类的属性的数据类型；

+++ Src/Visual Studio/SDK/C#/Eruru.ChatRobotRPC For Shared Project/Informations/ChatRobotFriendInformation.cs
@@ -20,7 +20,7 @@
 		/// <summary>
 		/// 年龄
 		/// </summary>
-		public long Age { get; set; }
+		public int Age { get; set; }
 		/// <summary>
 		/// 国家
 		/// </summary>
