+++ Src/Visual Studio/SDK/C#/Eruru.ChatRobotRPC For Shared Project/Informations/ChatRobotGroupMemberInformation.cs
@@ -20,11 +20,11 @@
 		/// <summary>
 		/// 群活跃积分
 		/// </summary>
-		public long ActiveIntegral { get; set; }
+		public int ActiveIntegral { get; set; }
 		/// <summary>
 		/// 群活跃等级
 		/// </summary>
-		public long ActiveLevel { get; set; }
+		public int ActiveLevel { get; set; }
 		/// <summary>
 		/// 加群时间戳（10位）
 		/// </summary>
@@ -36,7 +36,7 @@
 		/// <summary>
 		/// 禁言时间（距禁言结束时间,秒）
 		/// </summary>
-		public long BanSpeakSeconds { get; set; }
+		public int BanSpeakSeconds { get; set; }
 		/// <summary>
 		/// 是否是好友
 		/// </summary>
